DialogFlow Integration/requirements.txt
google-cloud-storage==1.16.0
dialogflow==0.6.0
Flask==1.0.2
Flask-RESTful==0.3.6